feat: define `quasispectrum` for non-unital algebras (#11219)
This defines the `quasispectrum` of an element `a : A` where `A` is a non-unital `R`-algebra. When `A` is unital and `R` is a semifield, then `quasispectrum R a = spectrum R a ∪ {0}`. The definition proceeds via *quasiregular* elements, which are also defined herein. The quasispectrum will serve as the domain of the continuous functions appearing in the continuous functional calculus for non-unital algebras.