fix(Cache): do not read lake-manifest.json at import-time (#11492)
Previously this file was read four times at runtime, and once for each file at compile time. Now it is read only once, when `lake exe cache` is run.

`initialize` is not a substitute for correctly passing global state through a program; my impression is that it is intended for *language-* rather than program-level initialization.

The performance impact is negligible, but the new code is more predictable.

Tested on mathlib, and on a package created using `lake new foo math`, pointed at this mathlib branch. `lake exe cache get` works correctly (leaves `lake build` with nothing to do) in both cases.